@@ -1,19 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Dapper.Core.Entities
+namespace Dapper.Core.Entities;
+
+public class Product
 {
-    public class Product
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public string Barcode { get; set; }
-        public decimal Rate { get; set; }
-        public DateTime AddedOn { get; set; }
-        public DateTime? ModifiedOn { get; set; }
-    }
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public string Description { get; set; }
+    public string Barcode { get; set; }
+    public decimal Rate { get; set; }
+    public DateTime AddedOn { get; set; }
+    public DateTime? ModifiedOn { get; set; }
 }
