@@ -1,58 +1,47 @@
 ﻿using Dapper.Application.Interfaces;
 using Dapper.Core.Entities;
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
-namespace Dapper.WebApi.Controllers
+namespace Dapper.WebApi.Controllers;
+
+[Route("api/[controller]")]
+[ApiController]
+public class ProductController(IUnitOfWork unitOfWork) : ControllerBase
 {
-    [Route("api/[controller]")]
-    [ApiController]
-    public class ProductController : ControllerBase
+    [HttpGet]
+    public async Task<IActionResult> GetAll()
     {
-        private readonly IUnitOfWork _unitOfWork;
+        var data = await unitOfWork.Products.GetAllAsync();
+        return Ok(data);
+    }
 
-        public ProductController(IUnitOfWork unitOfWork)
-        {
-            _unitOfWork = unitOfWork;
-        }
+    [HttpGet("{id}")]
+    public async Task<IActionResult> GetById(int id)
+    {
+        var data = await unitOfWork.Products.GetByIdAsync(id);
+        if (data == null) return Ok();
+        return Ok(data);
+    }
 
-        [HttpGet]
-        public async Task<IActionResult> GetAll()
-        {
-            var data = await _unitOfWork.Products.GetAllAsync();
-            return Ok(data);
-        }
+    [HttpPost]
+    public async Task<IActionResult> Add(Product product)
+    {
+        var data = await unitOfWork.Products.AddAsync(product);
+        return Ok(data);
+    }
 
-        [HttpGet("{id}")]
-        public async Task<IActionResult> GetById(int id)
-        {
-            var data = await _unitOfWork.Products.GetByIdAsync(id);
-            if (data == null) return Ok();
-            return Ok(data);
-        }
+    [HttpDelete]
+    public async Task<IActionResult> Delete(int id)
+    {
+        var data = await unitOfWork.Products.DeleteAsync(id);
+        return Ok(data);
+    }
 
-        [HttpPost]
-        public async Task<IActionResult> Add(Product product)
-        {
-            var data = await _unitOfWork.Products.AddAsync(product);
-            return Ok(data);
-        }
-
-        [HttpDelete]
-        public async Task<IActionResult> Delete(int id)
-        {
-            var data = await _unitOfWork.Products.DeleteAsync(id);
-            return Ok(data);
-        }
-
-        [HttpPut]
-        public async Task<IActionResult> Update(Product product)
-        {
-            var data = await _unitOfWork.Products.UpdateAsync(product);
-            return Ok(data);
-        }
+    [HttpPut]
+    public async Task<IActionResult> Update(Product product)
+    {
+        var data = await unitOfWork.Products.UpdateAsync(product);
+        return Ok(data);
     }
 }
